requirements.txt
django == 1.8.6
psycopg2
PyMySql
django-celery
Pillow
